Feature/version1.0.0

+++ ha-batch/docker/daily_analysis_api_log.sh
@@ -11,6 +11,7 @@
 # 
 # output: S3.analysis/YYYYMMDD/api_log.csv.gz
 # ----------------------------------------------------------------------------------------
+set -eu
 
 echo "------------------------------------------------------------------------"
 echo "START $0"
@@ -18,9 +19,6 @@ echo "------------------------------------------------------------------------"
 
 # call common.sh
 . ./common.sh
-
-# 処理対象年月YYYYMMDD
-DATE_OPTION_VALUE=$1
 
 # 結果チェック関数
 # 引数1：ジョブ名
@@ -33,7 +31,22 @@ check_result() {
     exit $result_code
   fi
 }
-# DATE_OPTION_VALUE="${DATE_OPTION_VALUE:-$(date +%Y%m%d)}"
+
+# ロック処理
+SCRIPT_NAME=$(basename "$0")
+LOCK_FILE_NAME="${SCRIPT_NAME}.lock"
+LOCK_FILE="${LOCK_DIR}/${LOCK_FILE_NAME}"
+mkdir -p "$(dirname "$LOCK_FILE")"
+
+exec 9>"${LOCK_FILE}"
+
+if ! flock -n 9; then
+  echo "[ERROR] dailyApiLogJob is already running."
+  exit 50
+fi
+
+# 処理対象年月YYYYMMDD
+DATE_OPTION_VALUE=$1
 
 cd ${BASE_DIR} && docker compose -f docker-compose.yml -f docker-compose.${ENV}.yml run --rm ha-batch --spring.batch.job.name=dailyApiLogJob d=${DATE_OPTION_VALUE}
 result=$?
